% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplexreg_links.R
\name{fixed_mean_links}
\alias{fixed_mean_links}
\alias{fixed_mean_link}
\alias{fixed_mean_link_inv}
\alias{fixed_mean_link_deriv1}
\alias{fixed_mean_link_deriv2}
\alias{fixed_mean_link_inv_deriv1}
\title{Fixed Mean Link Functions and Derivatives}
\usage{
fixed_mean_link(
  mu,
  type = c("logit", "probit", "loglog", "cloglog", "cauchit")
)

fixed_mean_link_inv(
  eta,
  type = c("logit", "probit", "loglog", "cloglog", "cauchit")
)

fixed_mean_link_deriv1(
  mu,
  type = c("logit", "probit", "loglog", "cloglog", "cauchit")
)

fixed_mean_link_deriv2(
  mu,
  type = c("logit", "probit", "loglog", "cloglog", "cauchit")
)

fixed_mean_link_inv_deriv1(
  eta,
  type = c("logit", "probit", "loglog", "cloglog", "cauchit")
)
}
\arguments{
\item{mu}{Mean parameter (numeric vector, \eqn{0 < \mu < 1}).}

\item{type}{Type of link function: \code{"logit"}, \code{"probit"},
\code{"loglog"}, \code{"cloglog"}, or \code{"cauchit"}.}

\item{eta}{Linear predictor of mean (numeric vector).}
}
\value{
A numeric vector corresponding to the evaluated link,
its inverse, or derivative depending on the function.
}
\description{
Provides the fixed mean link functions, their inverses, and derivatives
for the simplex regression model. Supported link types are:
\code{"logit"}, \code{"probit"}, \code{"loglog"}, \code{"cloglog"}, and \code{"cauchit"}.
}
\details{
Available link functions:
\itemize{
  \item Logit (\code{"logit"}): \eqn{g(\mu) = \log(\mu/(1-\mu))};
  \item Probit (\code{"probit"}): \eqn{g(\mu) = \Phi^{-1}(\mu)};
  \item Log-log (\code{"loglog"}): \eqn{g(\mu) = -\log(-\log(\mu))};
  \item Complementary log-log (\code{"cloglog"}): \eqn{g(\mu) = \log(-\log(1-\mu))};
  \item Cauchit (\code{"cauchit"}): \eqn{g(\mu) = \tan(\pi(\mu - 0.5))}.
}
}
\examples{
fixed_mean_link(0.5, type = "logit")
fixed_mean_link(c(0.2, 0.5, 0.8), type = "probit")
fixed_mean_link_inv(eta = 0.2, type = "logit")
fixed_mean_link_deriv1(mu = 0.5, type = "logit")

}
\seealso{
\code{\link{simplexreg}}, \code{\link{parametric_mean_links}},
\code{\link{dispersion_links}}.
}
