% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplexreg_links.R
\name{parametric_mean_links}
\alias{parametric_mean_links}
\alias{parametric_mean_link}
\alias{parametric_mean_link_inv}
\alias{parametric_mean_link_deriv1}
\alias{parametric_mean_link_inv_deriv1}
\alias{parametric_mean_link_deriv2}
\title{Parametric Mean Link Functions and Derivatives}
\usage{
parametric_mean_link(mu, lambda, type = c("plogit1", "plogit2"))

parametric_mean_link_inv(eta, lambda, type = c("plogit1", "plogit2"))

parametric_mean_link_deriv1(mu, lambda, type = c("plogit1", "plogit2"))

parametric_mean_link_inv_deriv1(eta, lambda, type = c("plogit1", "plogit2"))

parametric_mean_link_deriv2(mu, lambda, type = c("plogit1", "plogit2"))
}
\arguments{
\item{mu}{Mean parameter (numeric vector, \eqn{0 < \mu < 1}).}

\item{lambda}{Power parameter (numeric scalar, \eqn{\lambda > 0}).}

\item{type}{Type of link function: \code{"plogit1"} or \code{"plogit2"}.}

\item{eta}{Linear predictor of mean (numeric vector).}
}
\value{
Numeric vector with transformed values.
}
\description{
Provides the parametric mean link functions, their inverses, and derivatives
for the simplex regression models. Two parametric link types are supported:
\code{"plogit1"} and \code{"plogit2"}.
}
\details{
Two parametric mean link functions are available, as proposed by
Justino and Cribari-Neto (2026):
\itemize{
  \item Parametric logit type 1 (\code{"plogit1"}): \eqn{g(\mu; \lambda) = \log((1-\mu)^{-\lambda} - 1)};
  \item Parametric logit type 2 (\code{"plogit2"}): \eqn{g(\mu; \lambda) = \log(\mu^\lambda / (1 - \mu^\lambda))}.
}
Their inverses and derivatives with respect to \eqn{\mu} are also implemented.
}
\examples{
parametric_mean_link(0.2, lambda = 1.2, type = "plogit2")
parametric_mean_link(c(0.2, 0.5, 0.8), lambda = 1.5, type = "plogit1")
parametric_mean_link_inv(0, lambda = 1, type = "plogit2")
parametric_mean_link_deriv1(0.5, lambda = 1, type = "plogit2")
parametric_mean_link_inv_deriv1(0, lambda = 1, type = "plogit2")
parametric_mean_link_deriv2(0.5, lambda = 1, type = "plogit2")

}
\references{
Justino, M. E. C. and Cribari-Neto, F. (2026).
Simplex regression with a flexible logit link: Inference and application
to cross-country impunity data.
\emph{Applied Mathematical Modelling}, \bold{154}, 116713. \doi{10.1016/j.apm.2025.116713}
}
\seealso{
\code{\link{simplexreg}}, \code{\link{fixed_mean_links}},
\code{\link{dispersion_links}}.
}
