% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplexreg_methods.R
\name{simplexreg.methods}
\alias{simplexreg.methods}
\alias{print.simplexregression}
\alias{summary.simplexregression}
\alias{coef.simplexregression}
\alias{vcov.simplexregression}
\alias{logLik.simplexregression}
\alias{fitted.simplexregression}
\alias{predict.simplexregression}
\alias{nobs.simplexregression}
\alias{df.residual.simplexregression}
\alias{deviance.simplexregression}
\alias{formula.simplexregression}
\alias{terms.simplexregression}
\alias{model.frame.simplexregression}
\alias{model.matrix.simplexregression}
\alias{update.simplexregression}
\alias{simulate.simplexregression}
\alias{AIC.simplexregression}
\alias{BIC.simplexregression}
\alias{hatvalues.simplexregression}
\alias{cooks.distance.simplexregression}
\alias{bread.simplexregression}
\alias{estfun.simplexregression}
\alias{coeftest.simplexregression}
\alias{lrtest.simplexregression}
\alias{print.summary.simplexregression}
\alias{HQIC}
\alias{HQIC.simplexregression}
\title{Methods for simplexregression Objects}
\usage{
\method{print}{simplexregression}(x, digits = max(3, getOption("digits") - 3), ...)

\method{summary}{simplexregression}(object, ...)

\method{print}{summary.simplexregression}(x, digits = max(3, getOption("digits") - 3), ...)

\method{coef}{simplexregression}(object, model = c("full", "mean", "dispersion"), ...)

\method{vcov}{simplexregression}(object, model = c("full", "mean", "dispersion"), ...)

\method{logLik}{simplexregression}(object, ...)

\method{fitted}{simplexregression}(object, ...)

\method{predict}{simplexregression}(
  object,
  newdata = NULL,
  type = c("response", "link", "dispersion"),
  ...
)

\method{nobs}{simplexregression}(object, ...)

\method{df.residual}{simplexregression}(object, ...)

\method{deviance}{simplexregression}(object, ...)

\method{formula}{simplexregression}(x, ...)

\method{terms}{simplexregression}(x, model = c("mean", "dispersion"), ...)

\method{model.frame}{simplexregression}(formula, ...)

\method{model.matrix}{simplexregression}(object, model = c("mean", "dispersion"), ...)

\method{update}{simplexregression}(object, formula., ..., evaluate = TRUE)

\method{simulate}{simplexregression}(object, nsim = 1, seed = NULL, ...)

\method{AIC}{simplexregression}(object, ..., k = 2)

\method{BIC}{simplexregression}(object, ...)

HQIC(object, ...)

\method{HQIC}{simplexregression}(object, ...)

\method{hatvalues}{simplexregression}(model, ...)

\method{cooks.distance}{simplexregression}(model, type = c("pearson", "weighted"), ...)

\method{bread}{simplexregression}(x, ...)

\method{estfun}{simplexregression}(x, ...)

\method{coeftest}{simplexregression}(x, vcov. = NULL, df = Inf, ...)

\method{lrtest}{simplexregression}(object, ...)
}
\arguments{
\item{digits}{Number of digits to printing.}

\item{...}{Additional arguments.}

\item{object, x}{An object of class \code{simplexregression}.}

\item{model}{Character specifying for which component of the model
coefficients/covariance should be extracted.}

\item{newdata}{Optional data frame for prediction.}

\item{type}{Character indicating type of predictions: fitted means of the
response (default, "response"), corresponding linear
predictor ("link") or fitted dispersion parameter ("dispersion").}

\item{formula}{A model formula or terms object.}

\item{formula.}{Changes to the formula.}

\item{evaluate}{If true evaluate the new call else return the call.}

\item{nsim}{number of response vectors to simulate. Defaults to 1.}

\item{seed}{an object specifying if and how the random number generator
should be initialized.}

\item{k}{weight of the penalty term in AIC. Default is 2.}

\item{vcov.}{a specification of the covariance matrix of the estimated coefficients.}

\item{df}{the degrees of freedom to be used.}
}
\description{
Methods for extracting information from fitted simplex regression model objects
of class \code{simplexregression}.
}
\examples{
data(ReadingSkills, package = "SimplexRegression")
fit <- simplexreg(accuracy ~ dyslexia * iq | dyslexia + iq + I(iq^2),
                 data = ReadingSkills)

# Extract information
summary(fit)
coef(fit)
vcov(fit)
logLik(fit)
fitted(fit)
AIC(fit)
BIC(fit)
HQIC(fit)
hatvalues(fit)
cooks.distance(fit)

}
\seealso{
\code{\link{simplexreg}}.
}
